Squad working, saving state before removing process_answer fct
